requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
streamlit==1.38.0
pydantic==2.9.2
pydantic-settings==2.5.2
requests==2.32.3
gunicorn==22.0.0
